% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertions.R
\name{cuts_to_frequency}
\alias{cuts_to_frequency}
\title{cuts_to_frequency}
\usage{
cuts_to_frequency(insertions)
}
\description{
From a tibble of cuts, derive a tibble of region-wise
cutting frequencies for each strand. Frequency for each strand
is with respect to all cuts within each regions, regardless of strand.
}
