% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMotifs.R
\name{find_motif_instances}
\alias{find_motif_instances}
\title{find_motif_instances}
\usage{
find_motif_instances(object, pwms, genome = "BSgenome.Hsapiens.UCSC.hg19")
}
\description{
Convenience/wrapper function for finding motifs with motifmatchr.
}
