% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalImporters.R
\name{import_narrowpeak}
\alias{import_narrowpeak}
\title{Import narrowPeak file}
\usage{
import_narrowpeak(path)
}
\arguments{
\item{path}{Character corresponding to path to narrowPeak formatted file.}
}
\description{
#https://charlesjb.github.io/How_to_import_narrowPeak/
}
