% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertions.R
\name{plot_insertion_freq}
\alias{plot_insertion_freq}
\title{make a plot of the output from the  tn5cuts_to_freq function}
\usage{
plot_insertion_freq(freq_tbl)
}
\description{
make a plot of the output from the  tn5cuts_to_freq function
}
