% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertions.R
\name{tn5_cuts_in_regions}
\alias{tn5_cuts_in_regions}
\title{tn5_cuts_in_regions}
\usage{
tn5_cuts_in_regions(regions, bamfile, cores = 1, bai = bamfile)
}
\description{
Find tn5 insertions across multiple regions in a single sample.
}
