% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertions.R
\name{tn5cuts_to_freq}
\alias{tn5cuts_to_freq}
\title{tn5cuts_to_freq}
\usage{
tn5cuts_to_freq(insertions)
}
\description{
From a tibble of cuts, derive a tibble of region-wise
cutting frequencies for each strand. Frequency for each strand
is with respect to all cuts within each region, regardless of strand.
}
